# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "RCPCH Digital Growth API",
        "description": "Returns SDS and centiles for child growth measurements using growth references. Currently provides calculations based on the UK-WHO, Turner's Syndrome and Trisomy-21 references.",
        "version": "5.0.0"
    },
    "paths": {
...
        "/cdc/calculation": {
            "post": {
                "tags": [
                    "cdc"
                ],
                "summary": "Cdc Calculation",
                "description": "## CDC Centile and SDS Calculations\n\n* These are the 'standard' centiles for children in North America. It uses a hybrid of the WHO and CDC datasets.\n* Returns a single centile/SDS calculation for the selected `measurement_method`.\n* Gestational age correction will be applied automatically if appropriate according to the gestational age at birth data supplied.\n* Available `measurement_method`s are: `height`, `weight`, `bmi`, or `ofc` (OFC = occipitofrontal circumference = 'head circumference').\n* Note that BMI must be precalculated for the `bmi` function.\n* Dates will discard anything after first 'T' in `YYYY-MM-DDTHH:MM:SS.milliseconds+TZ` etc\n* Optional Bone age data associated with a height can be passed:\n*   - `bone_age` as a float in years\n*   - `bone_age_sds` and `bone_age_centile` as floats\n*   - `bone_age_type` as one of `greulich-pyle`, `tanner-whitehouse-ii`, `tanner-whitehouse-iiI`, `fels`, `bonexpert`\n* Optional events can be passed in as a list of strings - each list is associated with a measurement",
                "operationId": "cdc_calculation_cdc_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MeasurementRequest",
                                "examples": [
                                    {
                                        "birth_date": "2020-04-12",
                                        "observation_date": "2028-06-12",
                                        "observation_value": 115,
                                        "sex": "female",
                                        "gestation_weeks": 40,
                                        "gestation_days": 0,
                                        "measurement_method": "height",
                                        "bone_age": 10,
                                        "bone_age_centile": 98,
                                        "bone_age_sds": 2.0,
                                        "bone_age_text": "This bone age is advanced",
                                        "bone_age_type": "greulich-pyle",
                                        "events_text": [
                                            "Growth hormone start",
                                            "Growth Hormone Deficiency diagnosis"
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cdc/bulk-calculation": {
            "post": {
                "tags": [
                    "cdc"
                ],
                "summary": "Cdc Bulk Calculation",
                "operationId": "cdc_bulk_calculation_cdc_bulk_calculation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BulkMeasurementRequest",
                                "examples": [
                                    {
                                        "measurement_method": "height",
                                        "birth_date": "2020-04-12",
                                        "sex": "female",
                                        "gestation_weeks": 40,
                                        "gestation_days": 0,
                                        "observations": [
                                            {
                                                "observation_date": "2028-06-12",
                                                "observation_value": 115
                                            },
                                            {
                                                "observation_date": "2028-12-12",
                                                "observation_value": 130
                                            }
                                        ]
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BulkMeasurementObject"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cdc/chart-coordinates": {
            "post": {
                "tags": [
                    "cdc"
                ],
                "summary": "Cdc Chart Coordinates",
                "description": "## CDC Chart Coordinates data.\n\n* Returns coordinates for constructing the lines of a traditional growth chart, in JSON format\n* Requires a sex ('male' or 'female' lowercase) and a measurement_method ('height', 'weight' ,'bmi', 'ofc')\n* If custom centiles/sds collections (individually or as a collection) are required, accepts a list of float values (up to 15) as centile_format parameter\n* The is_sds boolean flag (default false) specifies if the custom list is of SDS or centiles.\n* In addition to the custom list, \"cole-nine-centiles\" or \"three-percent-centiles\" can be specified which are standard collections.\n* If no centile_format is supplied, \"cole-nine-centiles\" are returned as a default.",
                "operationId": "cdc_chart_coordinates_cdc_chart_coordinates_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChartCoordinateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centile_Data"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cdc/fictional-child-data": {
            "post": {
                "tags": [
                    "cdc"
                ],
                "summary": "Fictional Child Data",
                "description": "## CDC Fictional Child Data Endpoint\n\n* Generates synthetic data for demonstration or testing purposes",
                "operationId": "fictional_child_data_cdc_fictional_child_data_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FictionalChildRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/MeasurementObject"
                                    },
                                    "type": "array",
                                    "title": "Response Fictional Child Data Cdc Fictional Child Data Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Request validation or application-generated error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntityResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BulkMeasurementObject": {
                "properties": {
                    "results": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/MeasurementObject"
                                },
                                {
                                    "$ref": "#/components/schemas/APIErrorDetail"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Results"
                    }
                },
                "type": "object",
                "required": [
                    "results"
                ],
                "title": "BulkMeasurementObject"
            },
...
            "Centile_Data": {
                "properties": {
                    "centile_data": {
                        "items": {
                            "$ref": "#/components/schemas/ReferenceCreate"
                        },
                        "type": "array",
                        "title": "Centile Data"
                    }
                },
                "type": "object",
                "required": [
                    "centile_data"
                ],
                "title": "Centile_Data"
            },
...
            "EventsData": {
                "properties": {
                    "events_text": {
                        "anyOf": [
                            {
                                "items": {},
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Events Text"
                    }
                },
                "type": "object",
                "title": "EventsData"
            },
...
            "MeasurementObject": {
                "properties": {
                    "provenance": {
                        "$ref": "#/components/schemas/Provenance"
                    },
                    "birth_data": {
                        "$ref": "#/components/schemas/BirthData"
                    },
                    "measurement_dates": {
                        "$ref": "#/components/schemas/MeasurementDates"
                    },
                    "child_observation_value": {
                        "$ref": "#/components/schemas/ChildObservationValue"
                    },
                    "measurement_calculated_values": {
                        "$ref": "#/components/schemas/MeasurementCalculatedValues"
                    },
                    "plottable_data": {
                        "$ref": "#/components/schemas/PlottableData"
                    },
                    "bone_age": {
                        "$ref": "#/components/schemas/BoneAge"
                    },
                    "events_data": {
                        "$ref": "#/components/schemas/EventsData"
                    }
                },
                "type": "object",
                "required": [
                    "provenance",
                    "birth_data",
                    "measurement_dates",
                    "child_observation_value",
                    "measurement_calculated_values",
                    "plottable_data",
                    "bone_age",
                    "events_data"
                ],
                "title": "MeasurementObject"
            },
            "MeasurementRequest": {
                "properties": {
                    "observation_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Observation Date",
                        "description": "Date of the observation, in the format YYYY-MM-DD."
                    },
                    "observation_value": {
                        "type": "number",
                        "title": "Observation Value",
                        "description": "The value of the measurement supplied. This is supplied as a floating point number. All measurements should be supplied as **centimetres**, with the exception of Body Mass Index which is supplied as kilograms per metre squared (kg/m\u00b2)."
                    },
                    "bone_age": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bone Age",
                        "description": "Bone age in years. Age is paired with measurement taken at chronological age."
                    },
                    "bone_age_type": {
                        "anyOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "greulich-pyle",
                                    "tanner-whitehouse-ii",
                                    "tanner-whitehouse-iii",
                                    "fels",
                                    "bonexpert"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bone Age Type",
                        "description": "Method used to calculate bone age. Must be one of `'greulich-pyle`, `tanner-whitehouse-ii`, `tanner-whitehouse-iii`, `fels`,`bonexpert`"
                    },
                    "bone_age_sds": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bone Age Sds",
                        "description": "The SDS of the bone age based on reference tables."
                    },
                    "bone_age_centile": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bone Age Centile",
                        "description": "The centile for the bone age based on reference tables."
                    },
                    "bone_age_text": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Bone Age Text",
                        "description": "Any report or contextual information relating to the bone age."
                    },
                    "events_text": {
                        "anyOf": [
                            {
                                "items": {},
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Events Text",
                        "description": "A list of strings. Contextual text which are associated with each measurement."
                    },
                    "gestation_days": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 6.0,
                                "minimum": 0.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gestation Days",
                        "description": "The number of additional days _beyond the completed weeks of gestation_ at which the patient was born, passed as an integer. Supplying this data enables Gestational Age correction if the child was not born at term. If no gestational age is passed then term is assumed. IMPORTANT: See also the other parameter `gestation_weeks` - both are usually required.",
                        "default": 0
                    },
                    "gestation_weeks": {
                        "anyOf": [
                            {
                                "type": "integer",
                                "maximum": 44.0,
                                "minimum": 22.0
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Gestation Weeks",
                        "description": "The number of completed weeks of gestation at which the patient was born, passed as an integer. Supplying this data enables Gestational Age correction if the child was not born at term. If no gestational age is passed then 40 weeks (term) is assumed. **IMPORTANT: See also the other parameter `gestation_days` - both are usually required.**",
                        "default": 40
                    },
                    "measurement_method": {
                        "type": "string",
                        "enum": [
                            "height",
                            "weight",
                            "ofc",
                            "bmi"
                        ],
                        "title": "Measurement Method",
                        "description": "The type of measurement performed on the infant or child as a string which can be `height`, `weight`, `bmi` or `ofc`. The value of this measurement is supplied as the `observation_value` parameter. The measurements represent height **in centimetres**, weight *in kilograms**, body mass index **in kilograms/metre\u00b2** and occipitofrontal circumference (head circumference, OFC) **in centimetres**."
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "title": "Birth Date",
                        "description": "Date of birth of the patient, in the format YYYY-MM-DD"
                    },
                    "sex": {
                        "type": "string",
                        "enum": [
                            "male",
                            "female"
                        ],
                        "title": "Sex",
                        "description": "The sex of the patient, as a string value which can either be `male` or `female`. Abbreviations or alternatives are not accepted."
                    }
                },
                "type": "object",
                "required": [
                    "observation_date",
                    "observation_value",
                    "measurement_method",
                    "birth_date",
                    "sex"
                ],
                "title": "MeasurementRequest",
                "description": "This class definition creates a Python model which can be converted by FastAPI to openAPI3 schema.\nWe aim to specify all textual information, constraints, and validation here.\nIt all ends up in the openAPI documentation, automagically."
            },
...
            "ReferenceCreate": {
                "additionalProperties": {
                    "$ref": "#/components/schemas/Sex"
                },
                "type": "object",
                "title": "ReferenceCreate"
            },
...
            "UnprocessableEntityResponse": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/APIErrorDetail"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": [
                    "detail"
                ],
                "title": "UnprocessableEntityResponse"
            }
        }
    }
}